Changed setup.sh. The Arm toolchain only installs on x86.

+++ bin/setup.sh
@@ -21,13 +21,15 @@ install_system_tools() {
 }
 
 setup_toolchain() {
-        echo "Setup tool chain."
-        mkdir -p $BUILD_DIR
-        cd $BUILD_DIR
-        rm -Rf gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu
-        wget -O gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz "https://developer.arm.com/-/media/Files/downloads/gnu-a/9.2-2019.12/binrel/gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz?revision=61c3be5d-5175-4db6-9030-b565aae9f766&la=en&hash=0A37024B42028A9616F56A51C2D20755C5EBBCD7"
-        tar xvf gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz
-        rm gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz
+	if [[ $(uname -p) == "x86_64" ]]; then
+		echo "Setup tool chain."
+		mkdir -p $BUILD_DIR
+		cd $BUILD_DIR
+		rm -Rf gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu
+		wget -O gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz "https://developer.arm.com/-/media/Files/downloads/gnu-a/9.2-2019.12/binrel/gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz?revision=61c3be5d-5175-4db6-9030-b565aae9f766&la=en&hash=0A37024B42028A9616F56A51C2D20755C5EBBCD7"
+		tar xvf gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz
+		rm gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz
+	fi
 }
 
 cd $(dirname $0)
